+++ anxcloud/provider.go
@@ -56,6 +56,7 @@ func Provider(version string) *schema.Provider {
 			"anxcloud_nic_types":             dataSourceNICTypes(),
 			"anxcloud_core_location":         dataSourceCoreLocation(),
 			"anxcloud_core_locations":        dataSourceCoreLocations(),
+			"anxcloud_vlan":                  dataSourceVLAN(),
 			"anxcloud_vlans":                 dataSourceVLANs(),
 			"anxcloud_tags":                  dataSourceTags(),
 			"anxcloud_cpu_performance_types": dataSourceCPUPerformanceTypes(),