+++ CHANGELOG.md
@@ -17,6 +17,10 @@ If the change isn't user-facing but still relevant enough for a changelog entry,
 * (internal)? scope: short description (#pr, @author)
 -->
 
+## Added
+
+* `anxcloud_vlan` data source (#165, @nachtjasmin)
+
 ### Changed
 * resource/anxcloud_ip_address: reserve an available address based on filters (#163, @anx-mschaefer)