+++ CHANGELOG.md
@@ -17,8 +17,9 @@ If the change isn't user-facing but still relevant enough for a changelog entry,
 * (internal)? scope: short description (#pr, @author)
 -->
 
-## Added
+## [0.6.3] - 2024-06-25
 
+## Added
 * `anxcloud_vlan` data source (#165, @nachtjasmin)
 
 ### Changed